requirements.txt
matplotlib>=3.8
numpy>=1.26
opencv-python-headless>=4.10
pandas>=2.2
Pillow>=10.0
python-dotenv>=1.0
requests>=2.31

